메인페이지 불필요내용 삭제
